pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cplt"
version = "0.1.0"
description = "Plot timeline/Gantt-style ranges from CSV files in the terminal"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "wesley", email = "rademakerwesley@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "plotext>=5.2",
    "pillow>=12.1.1",
]

[project.urls]
Homepage = "https://github.com/Warhorze/cplt"
Repository = "https://github.com/Warhorze/cplt"
Issues = "https://github.com/Warhorze/cplt/issues"

[project.optional-dependencies]
dev = [
    "commitizen>=4.0",
    "pytest>=7.0",
    "ruff>=0.1",
    "pyright>=1.1",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
]

[project.scripts]
cplt = "cplt.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/cplt"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "pep621"
update_changelog_on_bump = true
changelog_file = "CHANGELOG.md"
tag_format = "v$version"

[tool.pytest.ini_options]
testpaths = ["tests"]
